Issue #19809: Remove Grid layout info banner in tabs tray
