+++ app/src/main/java/org/mozilla/fenix/tabstray/TabsTrayInfoBannerBinding.kt
@@ -49,39 +49,11 @@ class TabsTrayInfoBannerBinding(
     }
 
     private fun displayInfoBannerIfNeeded(settings: Settings) {
-        banner = displayGridViewBannerIfNeeded(settings)
-            ?: displayAutoCloseTabsBannerIfNeeded(settings)
+        banner = displayAutoCloseTabsBannerIfNeeded(settings)
 
         banner?.apply {
             infoBannerView.visibility = VISIBLE
             showBanner()
-        }
-    }
-
-    private fun displayGridViewBannerIfNeeded(settings: Settings): InfoBanner? {
-        return if (
-            settings.shouldShowGridViewBanner &&
-            settings.canShowCfr &&
-            settings.listTabView
-        ) {
-            InfoBanner(
-                context = context,
-                message = context.getString(R.string.tab_tray_grid_view_banner_message),
-                dismissText = context.getString(R.string.tab_tray_grid_view_banner_negative_button_text),
-                actionText = context.getString(R.string.tab_tray_grid_view_banner_positive_button_text),
-                container = infoBannerView,
-                dismissByHiding = true,
-                dismissAction = {
-                    metrics?.track(Event.TabsTrayCfrDismissed)
-                    settings.shouldShowGridViewBanner = false
-                }
-            ) {
-                navigationInteractor.onTabSettingsClicked()
-                metrics?.track(Event.TabsTrayCfrTapped)
-                settings.shouldShowGridViewBanner = false
-            }
-        } else {
-            null
         }
     }
